requirements.txt
py-cord
py-cord[voice]
pydub
python-dotenv